requirements.txt
Django==1.9
django-settings-export==1.1.0
django-tinymce==2.4.0
jsonfield==1.0.3
